+++ transforms/language/pii_redactor/dpk_pii_redactor/ray/__init__.py
@@ -1,2 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
-from .transform import *
+from .runtime import *
