+++ transforms/language/pii_redactor/kfp_ray/pii_redactor_wf.py
@@ -28,7 +28,7 @@ task_image = "quay.io/dataprep1/data-prep-kit/pii-redactor-ray:latest"
 S3_SECRET = "s3-secret"
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "-m dpk_pii_redactor.ray.transform"
+EXEC_SCRIPT_NAME: str = "-m dpk_pii_redactor.ray.runtime"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:latest"
